add base_groups
