+++ app/controllers/group_admins_controller.rb
@@ -1,15 +1,22 @@
-class GroupAdminsController < ApplicationController
+class GroupAdminsController < BaseGroupsController
 
   before_action :admin_required
 
-  def create
-    current_group.group_members.find_by(user_id: params[:id]).admin!
-    redirect_to :back, notice: "メンバーを管理者登録しました"
+  def update
+    if params[:act] == "registration"
+      current_group.group_members.find_by(user_id: params[:id]).admin!
+      redirect_to :back, notice: "メンバーを管理者登録しました"
+    else
+      current_group.group_members.find_by(user_id: params[:id]).operator!
+      redirect_to :back, notice: "メンバーの管理者権限を削除しました"
+    end
   end
 
-  def destroy
-    current_group.group_members.find_by(user_id: params[:id]).operator!
-    redirect_to :back, notice: "メンバーの管理者権限を削除しました"
+
+  private
+
+  def current_group
+    Group.find(params[:group_id])
   end
 
 end
