+++ app/helpers/base_groups_helper.rb
@@ -1,0 +1,2 @@
+module BaseGroupsHelper
+end
