+++ app/views/group_members/index.html.erb
@@ -7,9 +7,9 @@
 			<%= link_to 'メンバー登録解除', group_group_member_path(id: member.id), method: :delete%>
 			 |
 				<% if current_group.admin?(member)%>
-					<%= link_to '管理者登録解除', group_group_admins_path(id: member.id), method: :delete%>
+					<%= link_to '管理者登録解除', group_group_admins_path(id: member.id, act: 'deregistration'), method: :patch%>
 				<% else %>
-					<%= link_to '管理者登録', group_group_admins_path(id: member.id), method: :post%>
+					<%= link_to '管理者登録', group_group_admins_path(id: member.id, act: 'registration'), method: :patch%>
 				<% end %>
 		<% end %>
 	<% end %>
