+++ spec/controllers/base_groups_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe BaseGroupsController, type: :controller do
+
+end
